Conditional Statements in python
Summary
In this section, we learned about conditional statements in Python. We covered the if statement to execute code based on a condition, the else statement for alternative actions when the condition is false, and elif for multiple conditions. Additionally, we looked at nested if statements for checking multiple layers of conditions. These constructs allow you to control the flow of your program based on varying conditions.